Add freely available analytics
This is to better understand the traffic going to Photoview.

The analytics are done using Plausible for better privacy/anonymity,
besides that the analytics data is made freely available at
https://analytics.qpqp.dk/photoview.github.io

To clarify, analytics will never be added to any Photoview software that
is intended to be run on personal servers or devices, such as the main repository.