index.njk is updated to resolve #36 Link to IOS app goes to blank page (#37)
* index.njk is updated to resolve #36 Link to IOS app goes to blank page

index.njk is updated to resolve #36 Link to IOS app goes to blank page.

Link to the app on the AppStore has been replaced with the link to the source files of the app (https://github.com/photoview/photoview-ios) as suggested in the comments.

* Update index.njk

"Source files" is changed to "source code" as suggester by the pr reviewer

* Update index.njk (FR)

French localization of index.njk has been updated in accordance with changes done to EN documentation to mention that iOS app is no longer maintained or available on AppStore and to provide the link to iOS app source code instead.